+++ Andead.SmartHome.UnitOfWork/Extensions/WorkflowStepExtensions.cs
@@ -5,7 +5,7 @@ namespace Andead.SmartHome.UnitOfWork.Extensions
 {
     public static class WorkflowStepExtensions
     {
-        public static IQueryBuilder<WorkflowStep> ByWorkflowId(this IQueryBuilder<WorkflowStep> repository, int workflowId)
+        public static IQueryBuilder<WorkflowStep> ByWorkflowId(this IQueryBuilder<WorkflowStep> repository, long workflowId)
         {
             return repository.Condition(new Specification<WorkflowStep>(x => x.WorkflowId == workflowId));
         }
